# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{id}": {
      "get": {
        "operationId": "GroupController_get",
        "summary": "Get group details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedGroupDetailsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GroupController_update",
        "summary": "Update a group",
        "description": "Name, description, and terms can be updated anytime. Core fields (contributionAmount, frequency, maxSlots, serviceCharge, lateFee) can only be updated before the group starts.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedGroupDetailsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GroupController_delete",
        "summary": "Delete a NOT_STARTED group (admin only)",
        "description": "Deletes a NOT_STARTED group and all contributors. Hard delete — no audit trail needed.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "No longer needed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/groups/{groupId}/invitations": {
      "post": {
        "operationId": "GroupInviteController_invite",
        "summary": "Invite a user to the group",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedIdResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GroupInviteController_list",
        "summary": "List pending invites for group",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query to filter results",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order: asc or desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for cursor-based paging",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedInviteListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/{groupId}/invitations/contact": {
      "post": {
        "operationId": "GroupInviteController_inviteContact",
        "summary": "Invite a non-app user by contact info (sends OTP)",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedIdResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/{groupId}/invitations/join-link": {
      "post": {
        "operationId": "GroupInviteController_createJoinLink",
        "summary": "Create a join link for group",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedJoinLinkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GroupInviteController_getJoinLink",
        "summary": "Get active join link for group",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedJoinLinkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GroupInviteController_revokeLink",
        "summary": "Revoke a join link",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GroupInviteController_updateJoinLinkStatus",
        "summary": "Update join link status (pause or resume)",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PAUSED",
                      "ACTIVE"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Link status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/groups/{groupId}/invitations/proxy-registrations/init": {
      "post": {
        "operationId": "GroupInviteController_proxyRegisterInit",
        "summary": "Step 1 – Admin initiates proxy registration: sends OTP via SMS to the user",
        "description": "An OTP is sent to the provided phone number. The (non-tech-savvy) user reads the OTP back to the admin verbally, confirming consent before the admin proceeds to confirm.",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyRegisterInitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP dispatched",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "OTP sent to user phone",
                    "phone": "+2348012345678",
                    "expiresInSeconds": 600
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/{groupId}/invitations/proxy-registrations/confirm": {
      "post": {
        "operationId": "GroupInviteController_proxyRegisterConfirm",
        "summary": "Step 2 – Admin confirms OTP read back by user, completing proxy registration",
        "description": "Validates the OTP, creates the user account (role: PROXY, notificationChannel: SMS), provisions a wallet, adds the user to the group, and sends a temporary password via SMS.",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyRegisterConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proxy user created and added to group",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Proxy user registered and added to group",
                    "userId": "uuid",
                    "contributorId": "uuid",
                    "loginEmail": "proxy.2348012345678@internal.hajor.app"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invitations": {
      "get": {
        "operationId": "InviteActionController_listMine",
        "summary": "List my received invitations",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query to filter results",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order: asc or desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for cursor-based paging",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of invitations"
          }
        },
        "tags": [
          "Invitations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invitations/{id}": {
      "patch": {
        "operationId": "InviteActionController_updateStatus",
        "summary": "Accept or reject an invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACCEPTED",
                      "REJECTED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invites/{id}/onboard": {
      "post": {
        "operationId": "PublicInviteController_onboard",
        "summary": "Complete onboarding for an invited user (no auth)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invites"
        ]
      }
    },
    "/join/{token}": {
      "post": {
        "operationId": "JoinLinkController_consume",
        "summary": "Join a group using a join link token",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "The join link token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully joined the group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Hajor API",
    "description": "API documentation for Hajor fintech system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WrappedOkResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/OkResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
...
      "OnboardInviteDto": {
        "type": "object",
        "properties": {}
      },
...
      "WrappedGroupDetailsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/GroupDetailsResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
...
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "My Savings Group"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "example": "Updated description"
          },
          "terms": {
            "type": "string",
            "maxLength": 5000,
            "example": "Updated terms..."
          },
          "contributionAmount": {
            "type": "number",
            "minimum": 0,
            "example": 5000
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ],
            "example": "WEEKLY"
          },
          "maxSlots": {
            "type": "number",
            "minimum": 2,
            "example": 10
          },
          "serviceCharge": {
            "type": "number",
            "minimum": 0,
            "example": 0
          },
          "lateFee": {
            "type": "number",
            "minimum": 0,
            "example": 0
          }
        }
      },
...
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateContactInviteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "phone"
        ]
      },
...
    }
  }
}